agents/requirements.txt
web3
goat-sdk
python-dotenv